Create my final project
It's my final project of SheCodes Basics. It was made only from my imagination, without an example. There is interaction with users in different variations, depending on the answers. In one of the scenarios, there is an automatic transition to the page of the actually existing foundaition of Serhiy Prytula.